Seems to work now. Using Double floating points

+++ src/Types.hs
@@ -3,9 +3,9 @@ module Types where
 import Data.Map(Map)
 
 type Sentence = [String]
-type BiProbMap= Map (String,String) Float
-type WTProbMap = Map (String,String) Float
+type BiProbMap= Map (String,String) Double
+type WTProbMap = Map (String,String) Double
 type TagsMap = Map String Int
 type TaggedSentence = [(String,String)]
-type Scores = Map (String,String) Float
+type Scores = Map (String,String) Double
 type BackTrack = Map (String,String) String
